requirements.txt
pdf2docx
python-docx
docx2pdf
PyPDF2
Pillow
pandas
openpyxl
pdf2image
pdfplumber
pytesseract
groq
python-dotenv
streamlit
